move onboarding config models to a seperate app
