+++ trustpoint/pki/util/idevid.py
@@ -9,8 +9,8 @@ from typing import TYPE_CHECKING
 from cryptography import x509
 from cryptography.x509.verification import Criticality, ExtensionPolicy, PolicyBuilder, Store, VerificationError
 
-from devices.models import (
-    DeviceModel,
+from devices.models import DeviceModel
+from onboarding.models import (
     OnboardingConfigModel,
     OnboardingPkiProtocol,
     OnboardingProtocol,
