+++ trustpoint/request/message_responder/cmp.py
@@ -14,7 +14,7 @@ from pyasn1.type import tag, univ, useful  # type: ignore[import-untyped]
 from pyasn1_modules import rfc2459, rfc4210  # type: ignore[import-untyped]
 from trustpoint_core.oid import HashAlgorithm, HmacAlgorithm
 
-from devices.models import OnboardingStatus
+from onboarding.models import OnboardingStatus
 from request.message_responder.base import AbstractMessageResponder
 from request.operation_processor import LocalCaCmpSignatureProcessor
 from request.request_context import CmpBaseRequestContext, CmpCertificateRequestContext
