+++ trustpoint/request/tests/test_authentication_extended.py
@@ -346,7 +346,7 @@ class TestCompositeAuthentication:
             # Update the existing config
             onboarding_config = device.onboarding_config
             # Only set password if protocol supports it
-            from devices.models import OnboardingProtocol
+            from onboarding.models import OnboardingProtocol
             if onboarding_config.onboarding_protocol == OnboardingProtocol.EST_USERNAME_PASSWORD:
                 onboarding_config.est_password = 'onboarding-password-123'
                 onboarding_config.save()
@@ -361,7 +361,7 @@ class TestCompositeAuthentication:
                 assert context.device == device
         else:
             # Create a new onboarding config with EST_USERNAME_PASSWORD
-            from devices.models import OnboardingProtocol
+            from onboarding.models import OnboardingProtocol
             onboarding_config = OnboardingConfigModel()
             onboarding_config.onboarding_protocol = OnboardingProtocol.EST_USERNAME_PASSWORD
             onboarding_config.est_password = 'onboarding-password-123'
