+++ trustpoint/request/tests/test_message_responder.py
@@ -5,7 +5,8 @@ from unittest.mock import Mock
 
 import pytest
 
-from devices.models import OnboardingStatus
+from devices.models import 
+from onboarding.models import OnboardingStatus
 from request.message_responder.est import (
     EstCertificateMessageResponder,
     EstErrorMessageResponder,
